# backend/docs/swagger.yaml
basePath: /
definitions:
  entity.AssignRoleMenusDto:
    properties:
      id:
        description: 角色id
        type: integer
      menuIds:
        items:
          type: integer
        type: array
    required:
    - id
    - menuIds
    type: object
...
  entity.CreateRoleDto:
    properties:
      description:
        type: string
      roleKey:
        type: string
      roleName:
        type: string
      roleStatus:
        enum:
        - 1
        - 2
        type: integer
    required:
    - roleKey
    - roleName
    type: object
...
  entity.DeleteRoleDto:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  entity.GetRoleByIdDto:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  entity.GetRoleMenusDto:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  entity.UpdateRoleDto:
    properties:
      description:
        type: string
      id:
        type: integer
      roleKey:
        type: string
      roleName:
        type: string
      roleStatus:
        enum:
        - 1
        - 2
        type: integer
    required:
    - id
    type: object
  entity.UpdateRoleStatusDto:
    properties:
      id:
        type: integer
      newStatus:
        enum:
        - 1
        - 2
        type: integer
    required:
    - id
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 后台管理系统API接口文档
  title: go-admin 后台管理系统
  version: "1.0"
paths:
...
  /api/roleService/assignRoleMenus:
    post:
      consumes:
      - application/json
      description: 角色权限分配
      parameters:
      - description: 权限菜单列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.AssignRoleMenusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 角色权限分配
      tags:
      - 角色管理
  /api/roleService/createRole:
    post:
      consumes:
      - application/json
      description: 创建角色
      parameters:
      - description: 创建角色请求结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.CreateRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建角色
      tags:
      - 角色管理
  /api/roleService/deleteRole:
    post:
      consumes:
      - application/json
      description: 删除角色
      parameters:
      - description: 删除角色请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.DeleteRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除角色
      tags:
      - 角色管理
  /api/roleService/getRoleById:
    post:
      consumes:
      - application/json
      description: 根据id查询角色
      parameters:
      - description: 根据id查询角色请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.GetRoleByIdDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 根据id查询角色
      tags:
      - 角色管理
  /api/roleService/getRoleDropdown:
    get:
      consumes:
      - application/json
      description: 角色下拉列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 角色下拉列表
      tags:
      - 角色管理
  /api/roleService/getRoleList:
    get:
      consumes:
      - application/json
      description: 查询角色列表
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        type: integer
      - description: 角色状态
        in: query
        name: roleStatus
        type: integer
      - description: 角色名称
        in: query
        name: roleName
        type: string
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 结束时间
        in: query
        name: endTiem
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 查询角色列表
      tags:
      - 角色管理
  /api/roleService/getRoleMenus:
    post:
      consumes:
      - application/json
      description: 查询角色权限列表
      parameters:
      - description: 查询角色权限
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.GetRoleMenusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 查询角色权限列表
      tags:
      - 角色管理
  /api/roleService/updateRole:
    post:
      consumes:
      - application/json
      description: 修改角色
      parameters:
      - description: 修改角色请求结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改角色
      tags:
      - 角色管理
  /api/roleService/updateRoleStatus:
    post:
      consumes:
      - application/json
      description: 修改角色状态
      parameters:
      - description: 修改角色状态请求结构体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateRoleStatusDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改角色状态
      tags:
      - 角色管理
...
securityDefinitions:
  BearerAuth:
    description: JWT认证令牌，格式Bearer <token>
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"